requirements.txt
numpy == 1.23.2
onnx == 1.12.0
onnxruntime-gpu == 1.12.1
opencv-python == 4.5.5.64
